+++ ZMachineLib/Operations/Kind0/Verify.cs
@@ -5,7 +5,7 @@ namespace ZMachineLib.Operations.Kind0
     public sealed class Verify : ZMachineOperation
     {
         public Verify(ZMachine2 machine)
-            : base(Kind0OpCodes.Verify, machine)
+            : base((ushort)Kind0OpCodes.Verify, machine)
         {}
 
         public override void Execute(List<ushort> args) => Jump(true);
